Add auto-generated "." and ".." entries to directories
There are pros and cons to real or fake "." and ".." dentries, but we
decided to go with fake because we don't have to do a potentially slow
lookup of ".." in large directories. Userspace likes "." and ".." to
come first in readdir(), and we return entries in hash value order, so
reserve 0 and 1 in the hash output space for "." and "..". Then
autogenerate them in readdir() and lookup() using the parent inode
back reference. Also add an optional position argument to the debugfs
readdir() command to test correctness of handling readdir() position.

Signed-off-by: Valerie Aurora <val@versity.com>