Insert \n before cmd output when input is not a tty
The output of piped input commands used by the tests is causing the
output to be concatenated on the line unless the cmd output itself
contains newlines. This is because when input is a pipe, we're not
going to be outputting the cmd's implicit newline. This makes the
output of test scripts a little awkward to read as we're seeing test
output that is relevant pre-pended with several <> line headings. Use
`isatty()` to detect whether stdin is a pipe or not, and insert an
extra newline before outputting cmd output resolves this.

Signed-off-by: Auke Kok <auke.kok@versity.com>