Don't overwrite op.val_size on btree delete
When the btree operation delete flag is set, the operation's val_size
is overwritten with the current size of the btree item. This causes a
bug on a simultaneous delete/insert btree operation in which the old
value size controls how many bytes of the new value are copied into
the value, instead of the new value size. Since delete-only btree
operations ignore op.val_size, just don't overwrite the operation
val_size at all.

Signed-off-by: Valerie Aurora <val@versity.com>