Add ngnfs_dir_mkdir() and debugfs command
Add ngnfs_dir_mkdir() and corresponding debugfs command. Other file
systems report a non-zero i_size for empty directories, so we also
give empty directories a non-zero file size. In our case it is 5
bytes, to match our convention for the effect on directory i_size of
real dirents of name plus null terminator as applied to "." and "..".

Signed-off-by: Valerie Aurora <val@versity.com>