Temporary hack to remove cache state on client disconnect
This is a racy inefficient version of removing a client's block cache
state when it disconnects, but it allows the devd server to continue
functioning after a client disconnects. The correct version would
remove the cache state when it received a message from the mapd server
that the client was evicted from the cluster. Clients would also be
identified by something other than the IP/port of their connection.

Signed-off-by: Valerie Aurora <val@versity.com>