Add replace_item() to btree operations
Allow replacement of existing btree items. We reuse freed space if the
new item is equal to or shorter than the existing one, otherwise we
allocate from the tail. If the item doesn't fit in the tail, we don't
attempt to compact the block to make it fit, we just split it.

Signed-off-by: Valerie Aurora <val@versity.com>